+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - The Highlight Symbol draw behavior is configurable between "Indent only selected"(`CombineIndent`) and "indent all"(`Static`)
 - Add ability to set style for `widget::Indicator` and `widget::OrIndicators` (fixes #2)
 - Apply common style for indent area
+- Add option to customize indent area style
 
 ## 0.1.0